Struct fun_collections::FunMap

pub struct FunMap<K, V> { /* private fields */ }

Implementations

impl<K: Clone + Ord, V: Clone> FunMap<K, V>

pub fn new() -> Self

pub fn iter(&self) -> Iter<'_, K, V>

pub fn for_each_mut<F: FnMut(&K, &mut V)>(&mut self, f: F)

Applies a function to every key-value pair in the map.

The passed function must take a reference to the key type and a mutable reference to the value type. Any shared nodes in the tree are cloned, regardless of whether the contained values are mutated.

This method provides partial compensation for the absence of iter_mut().

Examples

use fun_collections::FunMap;

let mut fmap = FunMap::new();
fmap.insert(0, "a");
fmap.for_each_mut(|_, v| *v = "b");
assert_eq!(fmap.get(&0), Some(&"b"));

pub fn insert(&mut self, key: K, val: V) -> Option<V>

Inserts a key-value pair in the map.

Examples

use fun_collections::FunMap;

let mut fmap = FunMap::new();
fmap.insert(0, "a");
assert_eq!(fmap.get(&0), Some(&"a"));

pub fn remove<Q>(&mut self, key: &Q) -> Option<V>where
K: Borrow<Q> + Ord,
Q: Ord + ?Sized,

Removes a key from a map and returns the unmapped value.

Examples

use fun_collections::FunMap;

let mut fmap = FunMap::new();
fmap.insert(1, 2);
fmap.insert(2, 3);
assert_eq!(fmap.remove(&2), Some(3));
assert_eq!(fmap.remove(&2), None);

pub fn join_with(&mut self, key: K, val: V, rhs: FunMap<K, V>)

Join the RHS into this map.

Requires: self.last_key_value().map_or(true, |(x,)| x < key); rhs.first_key_value().map_or(true, |(y,)| key < y);

pub fn join(lhs: &Self, key: K, val: V, rhs: &Self) -> Self

Build a new map by joining two maps around a pivot key that divides the entries of the maps.

The constructed map contains of the entries from both maps and the pivot key and the value provided for the pivot key.

Requires: lhs.last_key_value().map_or(true, |(m,)| m < key); rhs.first_key_value().map_or(true, |(n,)| key < n);

Examples

use fun_collections::FunMap;

let f1 = FunMap::from([(0, 'a'), (1, 'b')]);
let f2 = FunMap::from([(3, 'd')]);
let f3 = FunMap::join(&f1, 2, 'c', &f2);
assert_eq!(f3.get(&0), Some(&'a'));
assert_eq!(f3.get(&2), Some(&'c'));
assert_eq!(f3.get(&3), Some(&'d'));

pub fn split_off<Q>(&mut self, key: &Q) -> (Option<(K, V)>, Self)where
K: Borrow<Q>,
Q: Ord,

Moves all elements greater than a key into a new map and returns it and the original key-value pair.

Examples

use fun_collections::FunMap;

let mut fmap: FunMap<_, _> = (0..10).map(|i| (i, i * 2)).collect();
let (orig_kv, higher_fives) = fmap.split_off(&5);
assert_eq!(orig_kv, Some((5, 10)));
assert_eq!(fmap.get(&4), Some(&8));
assert_eq!(fmap.get(&6), None);
assert_eq!(higher_fives.get(&6), Some(&12));

pub fn split<Q>(map: &Self, key: &Q) -> (Self, Option<(K, V)>, Self)where
K: Borrow<Q>,
Q: Ord,

Splits a map on a key returning one map with entries less than the key, one map with entries greater than the key, and the entry corresponding to the key.

Examples

use fun_collections::FunMap;

let fmap = FunMap::from([(0,1),(1,2),(2,3)]);
let (lt, kv, gt) = FunMap::split(&fmap, &1);
assert_eq!(kv, Some((1, 2)));
assert_eq!(lt.get(&0), Some(&1));
assert_eq!(gt.get(&2), Some(&3));

pub fn diff_with(&mut self, other: Self)

Removes entries with keys from the other map.

Examples

use fun_collections::FunMap;

let mut lhs = FunMap::from([(0,1), (1, 2)]);
let rhs = FunMap::from([(1,5), (3,4)]);
lhs.diff_with(rhs);
assert_eq!(lhs.get(&0), Some(&1));
assert_eq!(lhs.get(&1), None);

pub fn diff(lhs: &Self, rhs: &Self) -> Self

Builds a map with entries from the LHS map with keys that are not in the RHS map.

Examples

use fun_collections::FunMap;

let lhs = FunMap::from([(0,1), (1, 2)]);
let rhs = FunMap::from([(1,5), (3,4)]);
let d = FunMap::diff(&lhs, &rhs);
assert_eq!(d.get(&0), Some(&1));
assert_eq!(d.get(&1), None);

pub fn sym_diff_with(&mut self, other: Self)

Removes entries with keys from the other map.

Examples

use fun_collections::FunMap;

let mut lhs = FunMap::from([(0,1), (1, 2)]);
let rhs = FunMap::from([(1,5), (3,4)]);
lhs.sym_diff_with(rhs);
assert_eq!(lhs.get(&0), Some(&1));
assert_eq!(lhs.get(&1), None);
assert_eq!(lhs.get(&3), Some(&4));

pub fn sym_diff(lhs: &Self, rhs: &Self) -> Self

Builds a map with entries from the LHS map with keys that are not in the RHS map.

Examples

use fun_collections::FunMap;

let lhs = FunMap::from([(0,1), (1, 2)]);
let rhs = FunMap::from([(1,5), (3,4)]);
let d = FunMap::sym_diff(&lhs, &rhs);
assert_eq!(d.get(&0), Some(&1));
assert_eq!(d.get(&1), None);
assert_eq!(d.get(&3), Some(&4));

pub fn intersect_with(&mut self, other: Self)

Discard entries that do not have a key from the other map.

Examples

use fun_collections::FunMap;

let mut lhs = FunMap::from([(0,1), (1, 2)]);
let rhs = FunMap::from([(1,5), (3,4)]);
lhs.intersect_with(rhs);
assert_eq!(lhs.get(&0), None);
assert_eq!(lhs.get(&1), Some(&2));

pub fn intersect(lhs: &Self, rhs: &Self) -> Self

Creates a map with entries from the LHS that have keys in the RHS.

Examples

use fun_collections::FunMap;

let lhs = FunMap::from([(0,1), (1, 2)]);
let rhs = FunMap::from([(1,5), (3,4)]);
let i = FunMap::intersect(&lhs, &rhs);
assert_eq!(i.get(&0), None);
assert_eq!(i.get(&1), Some(&2));

pub fn union_with(&mut self, other: Self)

Adds the entries from other that don’t have keys in this map.

Examples

use fun_collections::FunMap;

let mut lhs = FunMap::from([(0,1), (1, 2)]);
let rhs = FunMap::from([(1,5), (3,4)]);
lhs.union_with(rhs);
assert_eq!(lhs.get(&0), Some(&1));
assert_eq!(lhs.get(&1), Some(&2));
assert_eq!(lhs.get(&3), Some(&4));

pub fn union(lhs: &Self, rhs: &Self) -> Self

Builds a map with entries from both maps, with entries from the LHS taking precedence when a key appears in both maps.

Examples

use fun_collections::FunMap;

let mut lhs = FunMap::from([(0,1), (1, 2)]);
let rhs = FunMap::from([(1,5), (3,4)]);
lhs.union_with(rhs);
assert_eq!(lhs.get(&0), Some(&1));
assert_eq!(lhs.get(&1), Some(&2));
assert_eq!(lhs.get(&3), Some(&4));

pub fn first_key_value(&self) -> Option<(&K, &V)>

Returns the key-value pair for the least key in the map

Examples

use fun_collections::FunMap;

let fmap = FunMap::from([(2,0), (1,0)]);
assert_eq!(fmap.first_key_value(), Some((&1, &0)));

pub fn last_key_value(&self) -> Option<(&K, &V)>

Returns the key-value pair for the greatest key in the map

Examples

use fun_collections::FunMap;

let fmap = FunMap::from([(2,0), (1,0)]);
assert_eq!(fmap.last_key_value(), Some((&2, &0)));

pub fn contains<Q>(&self, k: &Q) -> boolwhere
K: Borrow<Q> + Ord,
Q: Ord + ?Sized,

pub fn get<Q>(&self, k: &Q) -> Option<&V>where
K: Borrow<Q> + Ord,
Q: Ord + ?Sized,

Returns a reference to the value associated with k.

Example

use fun_collections::FunMap;

let mut fmap = FunMap::new();
fmap.insert(0, 100);

assert_eq!(fmap.get(&0), Some(&100));

pub fn get_mut<Q>(&mut self, k: &Q) -> Option<&mut V>where
K: Borrow<Q> + Ord,
Q: Ord + ?Sized,

Returns a reference to the value associated with k.

Example

use fun_collections::FunMap;

let mut fmap = FunMap::new();
fmap.insert(1, 7);

*fmap.get_mut(&1).unwrap() = 2;
assert_eq!(fmap.get(&1), Some(&2));

pub fn entry(&mut self, key: K) -> Entry<'_, K, V>

pub fn is_empty(&self) -> bool

pub fn len(&self) -> usize